[WIP] ORM error with flyway

+++ src/main/java/com/example/sondagecoincafe/bll/UserService.java
@@ -5,7 +5,7 @@ import com.example.sondagecoincafe.bo.User;
 import java.util.Optional;
 
 public interface UserService {
-    Optional<User> getById(int id);
+    Optional<User> getById(Long id);
     Optional<User> getByUsername(String username);
     void createUser(User user);
 }
